backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-ldap==3.4.4
pyjwt==2.10.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
